+++ angular/lib/angular.dart
@@ -28,3 +28,4 @@ export 'src/core/linker.dart';
 export 'src/core/testability/testability.dart';
 export 'src/platform/browser/tools/tools.dart'
     show enableDebugTools, disableDebugTools;
+export 'src/runtime/check_binding.dart' show debugCheckBindings;