+++ _goldens/test/_files/i18n/text.dart
@@ -3,7 +3,11 @@ import 'package:angular/angular.dart';
 @Component(
   selector: 'message',
   template: '''
-    <p @i18n="description" @i18n.meaning="meaning" @i18n.skip>
+    <p
+        @i18n="description"
+        @i18n.locale="en_US"
+        @i18n.meaning="meaning"
+        @i18n.skip>
       message
     </p>
   ''',