+++ _tests/test/compiler/schema/dom_element_schema_registry_test.dart
@@ -2,7 +2,7 @@
 import 'package:test/test.dart';
 import 'package:angular/src/compiler/schema/dom_element_schema_registry.dart'
     show DomElementSchemaRegistry;
-import 'package:angular/src/core/security.dart';
+import 'package:angular/src/compiler/security.dart';
 
 void main() {
   group('DOMElementSchema', () {